CleanroomContainer
